move e2e outside src folder
