+++ e2e/channels.spec.ts
@@ -1,4 +1,4 @@
-import { HoprSDK as SDK } from '../sdk';
+import { HoprSDK as SDK } from '../src/sdk';
 
 const { HOPRD_API_TOKEN, HOPRD_API_ENDPOINT_1, HOPRD_API_ENDPOINT_2 } =
   process.env;
