+++ package.json
@@ -34,8 +34,8 @@
     "pack": "tsup && yarn pack",
     "lint": "tsc",
     "format": "prettier --check \"src/**/*.{ts,md}\"",
-    "test": "jest src/api",
-    "test:e2e": "bash src/e2e/start.sh",
+    "test": "jest src/",
+    "test:e2e": "bash e2e/start.sh",
     "docs": "typedoc --plugin typedoc-plugin-markdown --plugin typedoc-plugin-zod --hideBreadcrumbs",
     "publish": "tsup & npm publish"
   },
