--- src/e2e/.env
@@ -1,11 +1,0 @@
-# The HOPR_PLUTO_VERSION is a variable used by HOPR PLUTO to use an specific GCP image
-HOPR_PLUTO_VERSION=1.93.7
-
-# API_TOKEN
-HOPRD_API_TOKEN="S3CR3T-T0K3N!"
-
-HOPRD_API_ENDPOINT_1="http://127.0.0.1:13301"
-HOPRD_API_ENDPOINT_2="http://127.0.0.1:13302"
-HOPRD_API_ENDPOINT_3="http://127.0.0.1:13303"
-HOPRD_API_ENDPOINT_4="http://127.0.0.1:13304"
-HOPRD_API_ENDPOINT_5="http://127.0.0.1:13305"
